parchis-pathfinder/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
pycairo>=1.20.0
PyGObject>=3.40.0
pillow>=10.0.0